requirements.txt
torch
torchvision
numpy
pandas
opencv-python
tqdm
PyYAML
scikit-learn
matplotlib
seaborn
